Update documentation about how to run the tests
The currently documented command

  bazel test --test_tag_filters=checks

neithers works when running it from the gerrit root dir nor when running
it from the checks plugin dir.

The new command

  bazel test //plugins/checks/...

works from the gerrit root dir and is the same command that is also used
for running the tests of other plugins (e.g. the tests for the
code-owners plugin are run by "bazel test //plugins/code-owners/...").

Change-Id: I50e2250872a4d01142d4ce0a4b645d7928ffbc48
Signed-off-by: Edwin Kempin <ekempin@google.com>